move search_handler.cc/h to database
