+++ src/database/search_handler.cc
