+++ src/database/search_handler.h
